requirements.txt
Django==1.7.6
docopt==0.4.0
mailchimp==2.0.9
psycopg2==2.6
recaptcha-client==1.0.6
requests==2.6.0
SQLAlchemy==0.9.9
